Get rid of indexes access from two threads

+++ src/modules/memory/task_memory_read.cpp
@@ -9,7 +9,7 @@ void task_memory_read::operation(uint64_t nb_ops)
 {
     assert(m_op_index < m_config.indexes->size());
     for (size_t i = 0; i < nb_ops; ++i) {
-        unsigned idx = m_config.indexes->at(m_op_index++);
+        uint64_t idx = m_config.indexes->at(m_op_index++);
         std::memcpy(m_scratch.ptr,
                     m_buffer + (idx * m_config.block_size),
                     m_config.block_size);
